move all the other components in app/components into one of the sub-folders
