+++ templates/skeleton/app/components/blocks/ProductImage.tsx
