+++ templates/skeleton/app/components/templates/CartMain/index.ts
@@ -1,0 +1,1 @@
+export {CartMain} from './CartMain';
