+++ templates/skeleton/app/components/templates/PageLayout/Header.tsx
@@ -2,7 +2,7 @@ import {Suspense} from 'react';
 import {Await, NavLink} from '@remix-run/react';
 import {type CartViewPayload, useAnalytics} from '@shopify/hydrogen';
 import type {HeaderQuery, CartApiQueryFragment} from 'storefrontapi.generated';
-import {useAside} from '~/components/Aside';
+import {useAside} from '~/components/sections/Aside';
 
 interface HeaderProps {
   header: HeaderQuery;
