+++ templates/skeleton/app/components/templates/PageLayout/index.ts
@@ -1,0 +1,1 @@
+export {PageLayout} from './PageLayout';
