+++ templates/skeleton/app/routes/products.$handle.tsx
@@ -9,9 +9,9 @@ import {
 } from '@shopify/hydrogen';
 import type {SelectedOption} from '@shopify/hydrogen/storefront-api-types';
 import {getVariantUrl} from '~/lib/variants';
-import {ProductPrice} from '~/components/ProductPrice';
-import {ProductImage} from '~/components/ProductImage';
-import {ProductForm} from '~/components/ProductForm';
+import {ProductPrice} from '~/components/blocks/ProductPrice';
+import {ProductImage} from '~/components/blocks/ProductImage';
+import {ProductForm} from '~/components/sections/ProductForm';
 
 export const meta: MetaFunction<typeof loader> = ({data}) => {
   return [{title: `Hydrogen | ${data?.product.title ?? ''}`}];
